requirements.txt
Django==4.2.3
django-admin-tools==0.9.2
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
Pillow==10.0.1



